% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplicar_restricciones.R
\name{aplicar_restricciones}
\alias{aplicar_restricciones}
\title{Aplicar restricciones}
\usage{
aplicar_restricciones(microdato_imputado, microdato_original, parametros)
}
\arguments{
\item{microdato_imputado}{Dataframe con microdato después de la imputación.}

\item{microdato_original}{Dataframe con microdato original con valores NA donde debiese imputarse.}

\item{parametros}{Dataframe con los valores mínimos y máximos para cada variable por año y mes.}
}
\value{
Microdato con los valores de nt, ro, re, og, ho y he truncados para no exceder los máximos y mínimos de las razones. Solo para cadenas que fueron sometidas a imputación.
}
\description{
Función que trunca los valores de una base de microdatos de acuerdo a valores máximos y mínimos para cada variable. Los parámetros de truncamiento aplican sobre nt y las razones (roho, hont, ognt, hent y rehe).
}
\examples{
\dontrun{
# Asumiendo que df_imputado es el microdato imptuado,
#df_original es el microdato previo a la imputación y
# ejemplo_parametros son los parámetros definidos para el periodo.
# df_truncado <-
  truncar_valores_imputados(
  microdato_imputado = df_imputado,
  microdato_original = df_original,
  parametros = ejemplo_parametros
)
}
}
