% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_razones.R
\name{calcular_razones}
\alias{calcular_razones}
\title{Calcular razones}
\usage{
calcular_razones(df, razon, dividendo, divisor)
}
\arguments{
\item{df}{Dataframe sobre el que se calcular치 la raz칩n.}

\item{razon}{El nombre de la columna con la razon resultante. Puede ser entregado sin comillas.}

\item{dividendo}{El nombre del dividendo de  la razon a calcular. Puede ser entregado sin comillas.}

\item{divisor}{El nombre del divisor de  la razon a calcular.. Puede ser entregado sin comillas.}
}
\value{
Mismo dataframe con una nueva columna nueva con el resultado de la raz칩n.
}
\description{
Funci칩n que calcula una columna con una razon en un data frame. Si  ambos valores son 0, resulta 0, si existe un divisor 0, da 0 (en vez de infinito).
}
\examples{
df <- tibble::tibble(
rol = 1,
ro = 2,
ho = 1
)
resultado <- calcular_razones(
  df = df,
  razon = roho,
  dividendo = ro,
  divisor = ho
)
}
