% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejemplo_microdato_mes_t1_na}
\alias{ejemplo_microdato_mes_t1_na}
\title{Ejemplo de microdato en el mes t-1, pero con valores NA.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31662 rows and 16 columns.
}
\usage{
ejemplo_microdato_mes_t1_na
}
\description{
Ejemplo de microdato en el mes t-1, pero con valores NA.
}
\keyword{datasets}
