% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejemplo_microdato_mes_t3}
\alias{ejemplo_microdato_mes_t3}
\title{Ejemplo de un microdato en el mes t-3.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31662 rows and 16 columns.
}
\usage{
ejemplo_microdato_mes_t3
}
\description{
Ejemplo de un microdato en el mes t-3.
}
\keyword{datasets}
