% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejemplo_parametros}
\alias{ejemplo_parametros}
\title{Un ejemplo de como debe ser el dataframe que define los parámetros para truncar variables que exceden mínimos y máximos.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 72 rows and 5 columns.
}
\usage{
ejemplo_parametros
}
\description{
Un ejemplo de como debe ser el dataframe que define los parámetros para truncar variables que exceden mínimos y máximos.
}
\keyword{datasets}
