% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejemplo_roles_comportamiento_unico}
\alias{ejemplo_roles_comportamiento_unico}
\title{Ejemplo de un vector con roles de comportamiento único para ser utilizado en las funciones del paquete.}
\format{
An object of class \code{numeric} of length 200.
}
\usage{
ejemplo_roles_comportamiento_unico
}
\description{
Ejemplo de un vector con roles de comportamiento único para ser utilizado en las funciones del paquete.
}
\keyword{datasets}
