% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excluir_comportamiento_unico.R
\name{excluir_comportamiento_unico}
\alias{excluir_comportamiento_unico}
\title{Crar variable comportamiento único}
\usage{
excluir_comportamiento_unico(microdato, roles_comportamiento_unico = NULL)
}
\arguments{
\item{microdato}{Dataframe con  los microdatos, debe tener la columna "rol".}

\item{roles_comportamiento_unico}{Vector opcional con los roles que deben ser considerados como comportamiento único, implicando que serán imputados directamente sin necesidad de buscar vecinos. Esto es especialmente útil para roles específicos con patrones de datos consistentes que no requieren comparación.}
}
\value{
Una lista de dos dataframes, el primero de ellos trae los roles sin comportamiento único, mientras que la segunda es un dataframe con los roles de comportamiento único.
}
\description{
Pequeña función que extrae los roles de comportamiento único de un dataframe.
}
\examples{
roles_comportamiento_unico <- seq(1,4)
microdato <-
  data.frame(
    rol = seq(1:10),
    nt = 1
  )
df <-
  excluir_comportamiento_unico(
  microdato = microdato,
  roles_comportamiento_unico = roles_comportamiento_unico
  )
}
