% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_ano_base.R
\name{imputar_ano_base}
\alias{imputar_ano_base}
\title{Imputación de Microdatos para el Año Base con Cambio de Año Base en la Encuesta de Remuneraciones}
\usage{
imputar_ano_base(df, roles_comportamiento_unico = NULL, n_cadenas_minimas = 4)
}
\arguments{
\item{df}{DataFrame que contiene los microdatos para todos los meses del año base, con una columna "mes" indicando el mes correspondiente a cada observación.}

\item{roles_comportamiento_unico}{Vector con los roles de comportamiento único.}

\item{n_cadenas_minimas}{Número de cadenas mínimas aceptadas para dar una cadena como corta como valida.}
}
\value{
Un DataFrame completo con los datos imputados para todos los meses del año base. Este resultado final representa un conjunto de datos coherente y listo para análisis, con valores faltantes adecuadamente reemplazados siguiendo la metodología descrita.
}
\description{
El proceso de imputación se lleva a cabo en varias etapas:

1. Iteración Forward: Desde febrero hasta diciembre, utilizando los datos del mes anterior como referencia para cada mes.

2. Iteración Backward: Desde noviembre hasta enero, también utilizando los datos del mes anterior como referencia.

3. Combinación de Resultados: Los valores imputados en ambas iteraciones se combinan, priorizando los de la iteración forward.

4. Iteración Forward con Datos Imputados: Se repite el proceso de febrero a diciembre, pero utilizando los datos ya imputados como referencia.

5. Iteración Backward con Datos Imputados: Similarmente, se repite desde noviembre hasta enero, utilizando los datos imputados como referencia.

6. Iteración forward con imptuacion por arrastre: Se imputa por arrastre desde febrero hasta diciembre en sentido forward. Se va utilizando el nuevo mes imptuado de forma sucesiva para imputar el nuevo mes.

7. Iteración backward con imputación por arrastre: Se imputa por arrastre desde noviembre hasta enero en sentido backward. Se va utilizando el nuevo mes imptuado de forma sucesiva para imputar el nuevo mes.


Este enfoque iterativo asegura una mejor adaptación de los datos imputados a las tendencias y variaciones dentro del año, considerando tanto la información más reciente como la precedente para ajustar los valores faltantes.
}
\details{
Realiza la imputación de los microdatos de la Encuesta de Remuneraciones para un año completo, utilizando un enfoque iterativo tanto forward (hacia adelante) como backward (hacia atrás), y luego combinando los resultados para mejorar la coherencia y calidad de los datos imputados. Este proceso es particularmente relevante para el cambio de año base 2023, asegurando que los datos anuales sean completos y representativos.
}
\examples{
\dontrun{
# Asumiendo que `df_anual` es tu DataFrame con los datos del año base:
df_imputado <- imputar_ano_base(df_anual)
}
}
