% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_mes.R
\name{imputar_mes}
\alias{imputar_mes}
\title{Imputación de Microdatos para la Encuesta de Remuneraciones Mensual}
\usage{
imputar_mes(
  microdato_t,
  microdato_t1,
  roles_comportamiento_unico = NULL,
  n_cadenas_minimas = 4
)
}
\arguments{
\item{microdato_t}{Dataframe con los microdatos del mes T, es decir, el periodo actual que necesita imputación.}

\item{microdato_t1}{Dataframe con los microdatos del mes T-1, utilizado como referencia inmediata para la imputación.}

\item{roles_comportamiento_unico}{Vector opcional con los roles que deben ser considerados como comportamiento único, implicando que serán imputados directamente sin necesidad de buscar vecinos. Esto es especialmente útil para roles específicos con patrones de datos consistentes que no requieren comparación.}

\item{n_cadenas_minimas}{Número entero que representa el número mínimo de observaciones que deben tener los vecinos para ser considerados válidos en el proceso de imputación. Por defecto, este valor es 4, asegurando que solo se utilicen vecinos con una cantidad mínima de datos para la imputación.}
}
\value{
Un dataframe con los microdatos del mes T imputados, incorporando valores para entradas previamente faltantes basados en los datos históricos y la metodología hot deck. Este dataframe resultante es esencial para análisis subsecuentes, asegurando que la información sobre remuneraciones sea completa y representativa.
}
\description{
Esta función realiza la imputación de microdatos para la encuesta de remuneraciones que se lleva a cabo cada mes, utilizando un método hot deck. La imputación se basa en los datos de hasta tres meses anteriores (T-1, T-2, T-3) para rellenar los valores faltantes en el mes actual (T), asegurando una mayor precisión y representatividad en los datos imputados.
}
\examples{
microdato_imputado_mes_t <-
  imputar_mes(
  microdato_t = ejemplo_microdato_mes_t_na,
  microdato_t1 = ejemplo_microdato_mes_t1_na,
  roles_comportamiento_unico = ejemplo_roles_comportamiento_unico,
  n_cadenas_minimas = 4
  )
}
