% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imptuar_mes_arrastre.R
\name{imputar_mes_arrastre}
\alias{imputar_mes_arrastre}
\title{Imputaci칩n de mes t  por arrastre}
\usage{
imputar_mes_arrastre(microdato_t, microdato_t1)
}
\arguments{
\item{microdato_t}{Dataframe con microdatos para el mes T.}

\item{microdato_t1}{Dataframe con microdatos para el mes T1.}
}
\value{
Dataframe del mes T imputado por arrastre con el mes T1.
}
\description{
Esta funci칩n imputa el mes t utilizando el mes t1 por medio de arrastre. Es decir, imputa NT, ROHO, HONT, OGNT, HENT y REHE con los valores de t1, posteriormente calcula  RO, RE, HO, HE OG, HT y CT a partir de los ratios arrastrados.  Si la cadena tiene NT = 0 (ya sea el valor original o el imputado), todos los dem치s valores se imputar치n por 0.
}
\examples{
microdato_imputado_mes_t <- imputar_mes_arrastre(microdato_t = ejemplo_microdato_mes_t_na,
                                                 microdato_t1 = ejemplo_microdato_mes_t1_na)
}
