% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_parametros_arrastre.R
\name{imputar_parametros_arrastre}
\alias{imputar_parametros_arrastre}
\title{Imputar parámetros por arrastre}
\usage{
imputar_parametros_arrastre(
  microdato_largo_mes_t,
  microdato_largo_mes_t1,
  variables_imputacion = c("nt", "roho", "hont", "ognt")
)
}
\arguments{
\item{microdato_largo_mes_t}{Df con microdato en formato largo para el mes t.}

\item{microdato_largo_mes_t1}{Df con microdato en formato largo para el mes t1.}

\item{variables_imputacion}{Vector con el nombre de las variables a  imputar, mantener valor por defecto: c("nt", "roho", "hont", "ognt").}
}
\value{
El microdato del mes t  con valores imputados por arrastre.
}
\description{
Imputa un dataframe con microdatos en formato largo por arrastre.
}
