% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/imputar_parametros_segun_nivel_representatividad.R
\name{imputar_parametros_segun_nivel_representatividad}
\alias{imputar_parametros_segun_nivel_representatividad}
\title{Imputar parámetros según nivel de representatividad}
\usage{
imputar_parametros_segun_nivel_representatividad(
  microdato_largo_mes_t,
  microdato_largo_mes_t1,
  df_nivel_representatividad,
  agregacion,
  variables_imputacion = c("nt", "roho", "hont", "ognt"),
  n_cadenas_minimas = 4
)
}
\arguments{
\item{microdato_largo_mes_t}{Df con microdato en formato largo para el mes t.}

\item{microdato_largo_mes_t1}{Df con microdato en formato largo para el mes t1.}

\item{df_nivel_representatividad}{Df con los niveles de representatividad y las variaciones dados para un nivel de agregación}

\item{agregacion}{Vector con el nombre de las columnas que definen el nivel de agregacion, por ejemplo: c("tamano", "categoria", "sexo", "grupo").}

\item{variables_imputacion}{Vector con el nombre de las variables a  imputar, mantener valor por defecto: c("nt", "roho", "hont", "ognt").}

\item{n_cadenas_minimas}{Número de cadenas mínimas aceptadas para dar una cadena como corta como valida.}
}
\value{
El microdato del mes t  con valores imptuados para el nivel de agregación en formato largo.
}
\description{
Imputa un dataframe con microdatos en formato largo a
}
\examples{

# Asumiendo que 'microdato_t' es un microdato para el mes t y
# 'microdato_t1' es un microdato para el mes t1.

# Establecer un nivel de representatividad.
tcsg <- c("tamano", "categoria", "sexo", "grupo")
# Establecer variable de interes
variable_interes = "nt"

# Calcular los indicadores básicos para dos microdatos (t y t1)
microdato_t <- calculador_indicadores_basicos(ejemplo_microdato_mes_t_na)
microdato_t1 <- calculador_indicadores_basicos(ejemplo_microdato_mes_t1_na)

# Calcular los niveles de repreentatividad
representatividad_nt <-
  calcular_representatividad_por_variable_y_agregacion(
  mes_t = microdato_t,
  mes_t1 = microdato_t1,
  variable_interes = variable_interes,
  agregacion = tcsg
  )
representatividad_nt

}
