% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_rehe.R
\name{imputar_rehe}
\alias{imputar_rehe}
\title{Imputación de la Variable "rehe" basada en Datos del Periodo Anterior}
\usage{
imputar_rehe(microdato_t, microdato_t1)
}
\arguments{
\item{microdato_t}{Dataframe con los microdatos del mes t.}

\item{microdato_t1}{Dataframe con los microdatos del mes t1.}
}
\value{
Devuelve el DataFrame original con la columna re imputada según los criterios establecidos.
}
\description{
Esta función imputa valores faltantes para la variable rehe (remuneraciones extraordinarias por horas extraordinarias) en un conjunto de datos.
}
\examples{
microdato_imputado_mes_t <-
  imputar_rehe(
    microdato_t = ejemplo_microdato_mes_t_na,
    microdato_t1 = ejemplo_microdato_mes_t1
  )
microdato_imputado_mes_t
}
