% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_variables.R
\name{imputar_variables}
\alias{imputar_variables}
\title{Imputación de Microdatos para la Encuesta de Remuneraciones Mensual}
\usage{
imputar_variables(
  microdato_t,
  microdato_t1,
  n_cadenas_minimas = 4,
  variables_imputacion
)
}
\arguments{
\item{microdato_t}{Dataframe con los microdatos del mes T.}

\item{microdato_t1}{Dataframe con los microdatos del mes T-1.}

\item{n_cadenas_minimas}{Número entero que representa el número mínimo de observaciones que deben tener los vecinos para ser considerados válidos en el proceso de imputación. Por defecto, este valor es 4, asegurando que solo se utilicen vecinos con una cantidad mínima de datos para la imputación.}

\item{variables_imputacion}{Es un vector con los nombres de las variables a imputar. Los valores admitidos son "nt", "roho", "hont" y "ognt".Los valores deben ser entregados como vector, por ejemplo: c("nt") o c("roho", "ognt"). Es importante tener en consideración que si se imputa más de una variable con esta función, los valores se imputaran de forma paralela. Es decir, por ejemplo si se considera c("nt", "hont"), los valores imputados de nt no servirán para estimar e imputar hont. En dicho caso, deben imputarse de forma sucesiva.}
}
\value{
Un dataframe con los microdatos del mes T imputados, incorporando valores para entradas previamente faltantes basados en los datos históricos y la metodología hot deck. Este dataframe resultante es esencial para análisis subsecuentes, asegurando que la información sobre remuneraciones sea completa y representativa.
}
\description{
Esta función imputa una o varias variables de la base de microdatos por medio de hot-deck. Puede usarse para imputar NT, RO, HO y OG.
NOTA: Si bien se pueden imputar varias variables a la vez, se recomienda hacerlo una por una para que los resultados de la imputación de NT sirvan para imputar HO y así sucesivamente.
}
\examples{
microdato_imputado_mes_t <- imputar_variables(microdato_t = ejemplo_microdato_mes_t_na,
                                       microdato_t1 = ejemplo_microdato_mes_t1_na,
                                       n_cadenas_minimas = 4,
                                       variables_imputacion = c("nt"))
}
