% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_mean_por_agregacion.R
\name{n_mean_por_agregacion}
\alias{n_mean_por_agregacion}
\title{Calculo número de observaciones y promedio para una variable de interés según un nivel de representatividad dado.}
\usage{
n_mean_por_agregacion(df, variable_interes, agregacion)
}
\arguments{
\item{df}{Dataframe que incluye los microdatos de un mes}

\item{variable_interes}{Cadena de texto que indica el nombre de la variable de interés (nombre de la columna)}

\item{agregacion}{Vector de caracteres con los nombres de las columnas que definen el nivel de desagregación deseado (por ejemplo, c("sexo", "grupo", "tamano", "categoria")).}
}
\value{
Un dataframe con el número de observaciones validas para una variable de interés y el promedio de esa variable, para un nivel de representatividad dado.
}
\description{
Calculo número de observaciones y promedio para una variable de interés según un nivel de representatividad dado. Esta función se utiliza para establecer el númer de cadenas observadas o esperadas para un nivel de desagregación y una variable de interés dada. En el contexto del cálculo, al usar sobre t1 se obtienen los valores "esperados", mientras que, al calcular para el mes t se obtienen los valores "observados".
}
\examples{
# Preparar la base
microdato_t <- calculador_indicadores_basicos(ejemplo_microdato_mes_t_na)
cadenas_esperadas_tcsg <-
  n_mean_por_agregacion(
  df = microdato_t,
  variable_interes = "nt",
  agregacion = c("tamano", "categoria", "sexo", "grupo")
  )
}
