% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_microdatos.R
\name{round_microdatos}
\alias{round_microdatos}
\title{Redondear microdatos según razones}
\usage{
round_microdatos(df, decimales_truncado = 6, decimales_redondeo = 0)
}
\arguments{
\item{df}{Dataframe con los microdatos  y las columnas correspondientes.}

\item{decimales_truncado}{A qué número de decimales se realizará el truncamiento (ceiling) de los valores previo al redondeo.}

\item{decimales_redondeo}{A qué número de decimales se hará el redondeo.}
}
\value{
Mismo dataframe de entrada con los valores redondeados.
}
\description{
Función que redondea los microdatos  para la base final. Todos los valores decimales de NT >0 y <1 se redondean a 1. Posteriormente se redondean OG, HO y HE manteniendo las razones de OGNT, HONT y HENT. Después se redondean RE y RO manteniendo la razón de  REHE y ROHO. Finalmente se recalcula ht y ct.
}
