% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformar_formato_largo.R
\name{transformar_formato_largo}
\alias{transformar_formato_largo}
\title{Transformar a formato largo}
\usage{
transformar_formato_largo(df)
}
\arguments{
\item{df}{Microdato con indicadores básicos calculados.}
}
\value{
Microdato en formato largo con las columnas variables y valor.
}
\description{
Transforma un microdato a formato largo.
}
\examples{
df <- calculador_indicadores_basicos(ejemplo_microdato_mes_t)
df <- transformar_formato_largo(df)
df
}
