% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncar_razon.R
\name{truncar_razon}
\alias{truncar_razon}
\title{Truncar razon}
\usage{
truncar_razon(df, var)
}
\arguments{
\item{df}{Microdato en formato largo con las columnas: ano, mes, rol, grupo, sexo, categoria, tamano, variable, valor, valor original, max y min}

\item{var}{String con el valor a truncar, por ejemplo "nt", "hont", etc.  Usar esta función para imputar nt y las razones básicas.}
}
\value{
Microdato con el valor truncado para la variable especificada.
}
\description{
Pequeña función que trunca un valor según los valores definidos en las columnas min y max, siempre y cuando el valor haya sido sometido a una imptuacion (es decir, el valor original sea NA).
}
\examples{
\dontrun{
# Asumiendo que microdato_largo es un data frame con microdatos en
# formato largo con las columnas requeridas.
microdato_largo_imputado <-
  truncar_razon(
  df = microdato_largo,
  var = "nt"
)
}
}
