% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unir_microdatos_segun_prioridad.R
\name{unir_microdatos_segun_prioridad}
\alias{unir_microdatos_segun_prioridad}
\title{Une dos microdatos priorizando el valor de un df por sobre el otro.}
\usage{
unir_microdatos_segun_prioridad(df_prioritario, df_secundario)
}
\arguments{
\item{df_prioritario}{Dataframme con los valores para el periodo t.}

\item{df_secundario}{Dataframme con los valores para el periodo t imputado por un segundo método con menos prioridad.}
}
\value{
Un dataframe con las mismas variables, pero priorizando el valor  existente en el df_prioritario. Si ese dataframe tiene un valor na, entonces utilizará el dato  obtenido del df_secundario.
}
\description{
Esta función une dos dataframes con microdatos para el mismo mes obtenidos a partir de dos métodos de imputación. Posteriormente, prefiere el valor del dataframe prioritario por sobre el valor del dataframe secundario para cada una de las variables de interés. En caso de que la variable sea NA para ambos periodos, se conserva el NA.
}
